+++ src/combinator/core.rs
@@ -177,7 +177,7 @@ where
         if input.eof_offset() == 0 {
             Ok(input.next_slice(0))
         } else {
-            Err(ErrMode::from_error_kind(input.clone(), ErrorKind::Eof))
+            Err(ErrMode::from_error_kind(input, ErrorKind::Eof))
         }
     })
     .parse_next(input)
@@ -209,7 +209,7 @@ where
     trace("not", move |input: &mut I| {
         let mut i = input.clone();
         match parser.parse_next(&mut i) {
-            Ok(_) => Err(ErrMode::from_error_kind(input.clone(), ErrorKind::Not)),
+            Ok(_) => Err(ErrMode::from_error_kind(input, ErrorKind::Not)),
             Err(ErrMode::Backtrack(_)) => Ok(()),
             Err(e) => Err(e),
         }
@@ -481,7 +481,7 @@ pub fn success<I: Stream, O: Clone, E: ParserError<I>>(val: O) -> impl Parser<I,
 #[doc(alias = "unexpected")]
 pub fn fail<I: Stream, O, E: ParserError<I>>(i: &mut I) -> PResult<O, E> {
     trace("fail", |i: &mut I| {
-        Err(ErrMode::from_error_kind(i.clone(), ErrorKind::Fail))
+        Err(ErrMode::from_error_kind(i, ErrorKind::Fail))
     })
     .parse_next(i)
 }