+++ tests/testsuite/custom_errors.rs
@@ -20,11 +20,11 @@ impl<'a> From<(&'a str, ErrorKind)> for CustomError {
 }
 
 impl<'a> ParserError<Partial<&'a str>> for CustomError {
-    fn from_error_kind(_: Partial<&'a str>, kind: ErrorKind) -> Self {
+    fn from_error_kind(_: &Partial<&'a str>, kind: ErrorKind) -> Self {
         CustomError(format!("error code was: {:?}", kind))
     }
 
-    fn append(self, _: Partial<&'a str>, kind: ErrorKind) -> Self {
+    fn append(self, _: &Partial<&'a str>, kind: ErrorKind) -> Self {
         CustomError(format!("{:?}\nerror code was: {:?}", self, kind))
     }
 }